+++ CHANGELOG.md
@@ -71,6 +71,9 @@
 * Add `EveOnline::ESI::Models::Alliance`
 * Add `EveOnline::ESI::Models::AllianceIcon`
 * Add `EveOnline::ESI::Models::CharacterPortrait`
+* Add `EveOnline::ESI::CharacterBlueprints`
+* Add `EveOnline::ESI::Models::Blueprint`
+* Drop `EveOnline::XML::CharacterBlueprints`. Use `EveOnline::ESI::CharacterBlueprints` instead.
 
 **v0.12.0**